refactor s3 handling
